Add cloud auth support to lsvd-server for disk replication
This change enables LSVD volumes and mounts to replicate data to
miren.cloud when cloud auth is configured. Key changes:

- Add CloudURL and PrivateKey fields to ServiceConfig for passing
  cloud auth credentials from runner to lsvd-server
- Update VolumeOps.InitLSVDVolume to support cloud replication and
  remote-only mode (no local storage, cloud only)
- Update MountOps.OpenLSVDDisk to use ReplicaWriter when cloud auth
  is available, or DiskAPISegmentAccess for remote-only volumes
- Add WithCloudAuth server option and create AuthClient in Run()
- Thread remoteOnly flag through volume and mount controllers
- Update runner.go to write cloud auth config to service config
- Update server_lsvd.go to load and pass cloud auth options

The implementation maintains backward compatibility - volumes work
without cloud auth configured, using local storage only. When cloud
auth is enabled, volumes can operate in three modes:
1. Local only (no cloud auth)
2. Replica mode (local + cloud replication)
3. Remote only (cloud storage only, no local)

Also includes debug CLI commands for inspecting lsvd-server state
and removes the embedded debug RPC interface in favor of outboard
process management.

+++ cli/commands/debug_disk.go
@@ -19,7 +19,7 @@ import (
 func DebugDiskCreate(ctx *Context, opts struct {
 	ConfigCentric
 	Name       string `short:"n" long:"name" description:"Name for the disk" required:"true"`
-	Size       int64  `short:"s" long:"size" description:"Size of disk in GB" default:"10"`
+	Size       int    `short:"s" long:"size" description:"Size of disk in GB" default:"10"`
 	Filesystem string `short:"f" long:"filesystem" description:"Filesystem type (ext4, xfs, btrfs)" default:"ext4"`
 	CreatedBy  string `short:"c" long:"created-by" description:"Creator ID for the disk"`
 	RemoteOnly bool   `short:"r" long:"remote-only" description:"Store disk only in remote storage (no local replica)"`
@@ -54,7 +54,7 @@ func DebugDiskCreate(ctx *Context, opts struct {
 	// Create disk entity
 	disk := &storage_v1alpha.Disk{
 		Name:       opts.Name,
-		SizeGb:     opts.Size,
+		SizeGb:     int64(opts.Size),
 		Filesystem: fs,
 		Status:     storage_v1alpha.PROVISIONING,
 		RemoteOnly: opts.RemoteOnly,
